requirements.txt
customtkinter
torch
numpy
torchvision
pillow